1. RESTFul. Описание архитектуры [#6884] - Data presence check moved to service layer.

+++ src/main/java/ru/job4j/auth/service/SimplePersonService.java
@@ -3,12 +3,15 @@ package ru.job4j.auth.service;
 import java.util.List;
 import java.util.Optional;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import ru.job4j.auth.model.Person;
 import ru.job4j.auth.repository.PersonRepository;
 
 @Service
 @AllArgsConstructor
+@Slf4j
 public class SimplePersonService implements PersonService {
 
     private final PersonRepository personRepository;
@@ -24,15 +27,26 @@ public class SimplePersonService implements PersonService {
     }
 
     @Override
-    public Person save(Person person) {
-        return personRepository.save(person);
+    public boolean save(Person person) {
+        boolean result = false;
+        try {
+            personRepository.save(person);
+            result = true;
+        } catch (DataAccessException | IllegalArgumentException e) {
+            log.error(e.getMessage(), e);
+        }
+        return result;
     }
 
     @Override
-    public void deletePerson(int id) {
-        Person person = new Person();
-        person.setId(id);
-        personRepository.delete(person);
+    public boolean deletePerson(int id) {
+        boolean isExists = personRepository.existsById(id);
+        if (isExists) {
+            Person person = new Person();
+            person.setId(id);
+            personRepository.delete(person);
+        }
+        return isExists;
     }
 
 }
